LPS-61474 move to internal package
