+++ modules/portal-security/portal-security-sso-cas/src/main/java/com/liferay/portal/security/sso/cas/internal/SSOImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.cas;
+package com.liferay.portal.security.sso.cas.internal;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
