+++ modules/portal-security/portal-security-sso-cas/src/main/java/com/liferay/portal/security/sso/cas/internal/portal/settings/web/portlet/action/PortalSettingsCASFormMVCActionCommand.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.cas.portal.settings.web.portlet.action;
+package com.liferay.portal.security.sso.cas.internal.portal.settings.web.portlet.action;
 
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseFormMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
