+++ modules/portal-security/portal-security-sso-cas/src/main/java/com/liferay/portal/security/sso/cas/internal/portal/settings/web/portlet/resource/bundle/PortalSettingsPortletResourceBundlePublisher.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.cas.portal.settings.web.portlet.resource.bundle;
+package com.liferay.portal.security.sso.cas.internal.portal.settings.web.portlet.resource.bundle;
 
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.settings.web.constants.PortalSettingsPortletKeys;
